reworking industry reprocess to take portionsize into account.

+++ model/sde/SDE-Industry/src/main/java/fr/guiguilechat/jcelechat/model/sde/industry/IndustryUsage.java
@@ -66,6 +66,6 @@ public class IndustryUsage {
 
 	public LinkedHashSet<String> materialTE = new LinkedHashSet<>();
 
-	public LinkedHashMap<String, Integer> reprocess = new LinkedHashMap<>();
+	public LinkedHashMap<String, Double> reprocess = new LinkedHashMap<>();
 
 }
